+++ src/autocv/use_cases/__init__.py
@@ -1,2 +1,19 @@
 """Application use cases."""
 
+from autocv.use_cases.bootstrap_workspace import BootstrapWorkspace, WorkspaceBootstrapResult
+from autocv.use_cases.v1_application_service import (
+    FreelanceOpportunityDraft,
+    JobApplicationDraft,
+    MissingDocumentSourceError,
+    V1ApplicationService,
+)
+
+__all__ = [
+    "BootstrapWorkspace",
+    "FreelanceOpportunityDraft",
+    "JobApplicationDraft",
+    "MissingDocumentSourceError",
+    "V1ApplicationService",
+    "WorkspaceBootstrapResult",
+]
+